test(rebuild): tighten provider recovery coverage
