+++ src/lib/onboard/machine/handlers/provider-inference-recovery.ts
@@ -22,7 +22,13 @@ interface ProviderRecoverySetupOptions {
 }
 
 interface ProviderRecoveryOptions {
-  /** The rebuild preflight already validated and rewrote this matching session identity. */
+  /**
+   * Rebuild recreate replaces the durable session after deleting the old sandbox, so its sandbox
+   * step must remain incomplete until creation succeeds. The locked rebuild pipeline validates the
+   * target before deletion, then writes that exact identity into the pending session before onboard.
+   * Remove this exception once the handoff carries a dedicated provider-recovery authorization
+   * receipt.
+   */
   authoritativeResumeConfig?: boolean;
 }
